Fix check_seapp segfault and undefined linking err
When LINK_SEPOL_STATIC was not defined, log_warning
attempt to be resolved by the linker. That symbol was
not defined as it should have been log_warn.

When a key would be validated in key_map_validate, a
key that would go unchecked, like user, could cuase a
segfault when the se_key was getting free'd no matter what
at the end.

Change-Id: If334ba7350e6d2ad1fa9bed142bb2fabe7caa057