Add better per-device sepolicy support.
This is a rewrite of the existing implementation.
Three new variables are now needed to add/modify
the exisitng base policy. They are, PRODUCT_SEPOLICY_REPLACE
and PRODUCT_SEPOLICY_UNION which govern what files
are replaced and concatenated, and PRODUCT_SEPOLICY_DIRS
which lists the various directories that will contain
the PRODUCT_SEPOLICY_REPLACE and PRODUCT_SEPOLICY_UNION
policy files.

Change-Id: Id33381268cef03245c56bc5242fec7da9b6c6493
Signed-off-by: rpcraig <rpcraig@tycho.ncsc.mil>