3rd party VPN applications use a tun device
Tun device is created by system and read/written to by the VPN app.
This adds a boolean, app_vpn, to allow untrusted_app to act as a VPN client
and vpn_app to use if a site wants to lock down 3rd party VPN access to a
single, approved client.

As an example there is a client signature added to mac_permissions.xml.

Change-Id: I41c2c0d4c36029fb79cc28af3df252c285326e7c