Introduce untrustedappdomain attribute
Introduce attribute untrustedappdomain
and macro untrustedapp_domain(domain)
to enable new domains off of a common
set. For instance, OEMs may have apps
that are essentially 3rd party apps
, as they are not signed with the
platform keys, but may make use of
platform extensions that should not be
allowed to market applications.

The existing approach would require
a large copy/paste of the existing
domain, and a need to re-create it
for the new type. This avoids this
as the new type can be passed to
untrustedapp_domain(new_type) and
allow rules use untrsutedappdomain
attribute.

Change-Id: I7028115767bf00a347abd8e9e080ea45ef5ec819
Signed-off-by: William Roberts <w.roberts@sta.samsung.com>