Revert "Add a key directory argument to insertkeys.py"
This reverts commit 020b5ff6311044ef7a2200dd4db69f5cccf46213.

This commit conflicted with:
Change-Id: If4f169d9ed4f37b6ebd062508de058f3baeafead

The tool sets is CWD to the TOP of the Android tree. All
paths in the scripts can then be relative to the TOP of the
tree. Also, with the newest ability to use environment variables
exporting DEFAULT_SYSTEM_DEV_CERTIFICATE and using it as an
environment variable in the reference keys.conf will yield
similar results.

Also this change would require all certificates to live in a
common directory, which may not be realistic.

Signed-off-by: William Roberts <wroberts@tresys.com>