Label /data/system/ndebugsocket with its own type.
Otherwise it defaults to the label of /data/system and
cannot be distinguished from any other socket in that directory.

If/when debuggerd is confined, it will require:
unix_socket_connect(debuggerd, system_ndebug, system_server)
This is already in our policy.

If/when system_server is confined, it will require:
allow system_server system_ndebug_socket:sock_file create_file_perms;
This is likewise already in our policy.

I could add those rules as part of this change now, but the
current pattern in AOSP policy seems to be to not allow anything
already allowed via unconfined_domain() in the policy.

Change-Id: I57179aa18786bd56d247f397347e546cca978e41
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>