Merge Samsung's mediaserver.te into AOSP.
Add allow rules to access binder, property service, socket, and directories and files.
Note that there are a few lines of rules are commented, such as those related to system, dumplog and drmserver_socket. The main reason is because the dependency has not met. Once the dependency is met, we may consider to uncomment them.

Change-Id: I1286bcf0309ea1656cdd7989ec488b4948a4af94