Confine shell domain, but leave it permissive for now.
Confine the domain for an adb shell, leaving it permissive until
sufficiently tested.  init_shell (shell spawned by init, e.g.
console service) is not altered by this change and remains
unconfined; we have it confined but that can be uploaded separately.
Introduce a shelldomain attribute for rules common to both domains.

Change-Id: I01ee2c7ef80b61a9db151abe182ef9af7623c461
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>