[DO NOT SUBMIT] Make bluetooth, nfc, radio and shell adb-installable
bluetooth, nfc, radio and shell are not explicitly declared
in installd.te. This prevents applications in those group
from upgrading by "adb install -r".

You can reproduce the issue by following step:
 1. adb pull /system/priv-app/Shell.apk
 2. adb install -r Shell.apk

This operation fails only if the app belongs to any of the
groups specified in the commit title.

We'd like to make sure if this is intended behavior or just
forgoten to be added when enebling SELinux.

Change-Id: Ic5009f415f5e415bfc4c916c5e93ca1acfccc6e6