Define a domain for the bootanim service.
Leave the domain permissive initially until it gets more testing.

There is an attempt to set a property but this is denied by DAC
since it runs in the graphics UID.  Unclear as to whether this should
be allowed or dontaudit'd.  dmesg output from init is:
  init: sys_prop: permission denied uid:1003  name:service.bootanim.exit

We presently allow the connection to the property service in policy
since we would otherwise get avc denials but do not allow setting
the property as the DAC check by init denies access before we reach
the SELinux check by init.

Change-Id: I9d88d76d1ffdc79a2eff4545d37a9e615482df50
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>