allow setting all properties on userdebug / eng builds
In 712ca0a4d5c3ff77179da2544aafd6eb8e5a70c2, shell
was refactored into user and non-user rules. As part of that
change, the non-user shell SELinux domain was removed from
unconfined and made permissive.

Normally a permissive domain is allowed to do anything, but that's
not true for userspace SELinux enforcement, such as the property
service. Even for permissive domains, we need to allow writes
to the property service.

Change-Id: I01ffaa781cd3ce794f6d78538b93b2d412477364