Provide defaults for /data/data and /data/user labeling.
To ensure that we do not label package directories with system_data_file
if the package is missing from packages.list or something else goes
wrong with the normal labeling based on seapp_contexts and packages.list.
This at least ensures that any such package directories will be labeled
with app_data_file instead.

Change-Id: I66f18695d596e55fb3d9a09633e2b2bb93d6965c
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>