Adding permission which is needed to execute a certain application
When I executed the following application, I found a violation log.

Application name: "큐라레: 마법도서관"
(You can download this application if you input "curare" in the search text box of Google Play Store)

Reproduction path:
1. Downloading the application
2. Executing the application
3. A denial log shows up

[ 6969.565785 / 03-28 18:38:25.517][0] type=1400 audit(1395999505.517:10): avc:  denied  { ptrace } for  pid=11393 comm=".korproc.google" scontext=u:r:zygote:s0 tcontext=u:r:untrusted_app:s0 tclass=process

If we can not merge the allow rule, I think that we had better find and remove the path that application runs the ptrace operation in the framework.

Change-Id: Ia6cb76d11f287fd82715ad05fe0fdd05ba9b00e9