Adding permission which is needed to execute a certain application
When I executed the following application, I found a violation log.

Application name: "큐라레: 마법도서관"
(You can download this application if you input "curare" in the search text box of Google Play Store)

Reproduction path:
1. Downloading the application
2. Executing the application

If we can not merge the allow rule, we should find and remove the path that application runs the ptrace operation in the framework.

Change-Id: Ia6cb76d11f287fd82715ad05fe0fdd05ba9b00e9