Restrict requesting contexts other than policy-defined defaults.
Writing to the /proc/self/attr files (encapsulated by the libselinux
set*con functions) enables a program to request a specific security
context for various operations instead of the policy-defined defaults.
The security context specified using these calls is checked by an
operation-specific permission, e.g. dyntransition for setcon,
transition for setexeccon, create for setfscreatecon or
setsockcreatecon, but the ability to request a context at all
is controlled by a process permission.  Omit these permissions from
domain.te and only add them back where required so that only specific
domains can even request a context other than the default defined by
the policy.

As an example of where this can make a difference, consider the
ability to create unlabeled files.  If a domain is allowed
create to unlabeled:file and is not allowed self:process setfscreate,
then it can only create unlabeled files where the kernel and policy
would by default assign unlabeled to the new file (e.g. when the
parent directory is unlabeled or when there is an explicit
type_transition rule that would assign unlabeled to the file).  If
however a domain is allowed create to unlabeled:file and is also
allowed self:process setfscreate, then the process can request the
unlabeled type for any file it creates, even in a labeled parent
directory, as long as it also has write and add_name to the parent
directory type.

Change-Id: I6a2fb1279318625a80f3ea8e3f0932bdbe6df676
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>