Restrict use of context= mount options.
Prior to this change, the init and recovery domains were
allowed unrestricted use of context= mount options to force
all files within a given filesystem to be treated as having a
security context specified at mount time.  The context= mount
option can be used in device-specific fstab.<board> files
to assign a context to filesystems that do not support labeling
such as vfat where the default label of sdcard_external is not
appropriate (e.g. /firmware on hammerhead).

Remove the ability to use context= mount options on filesystems
that support labeling (e.g. ext4) so that they cannot be used
to remap e.g. /system or /data to a different type.  Also
remove the ability to remap any filesystem to the unlabeled type
via context= mount options.  Reinforce these restrictions via
neverallow rules on all domains.

Change-Id: I09e13839b1956f61875a38844fe4fc3c911ea60f
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>