Remove /data/security access from unconfineddomain.
Ensure that only init and system_server can set the selinux.reload_policy
property to trigger a policy reload, and that only the system_server can
write to /data/security to install the policy files.  The init.rc file
sets the property from post-fs-data to trigger a reload once /data is
mounted, and the SELinuxPolicyInstallReceiver writes policy updates to
/data/security and sets the property to trigger a reload when an update
bundle is received.

This is only truly meaningful if the support for /data/security policies
is restored, but is harmless otherwise.

Change-Id: I5ad5e3b6fc7abaafd314d31723f37b708d8fcf89
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>